+++ src/router.tsx
@@ -7,8 +7,8 @@ import ForumPostBoard from '@components/ForumPost/ForumPostBoard';
 import React from 'react';
 import { Navigate, Outlet, createBrowserRouter } from 'react-router-dom';
 
-import { ConsultingHistory } from '@/components/ConsultingManagement/ConsultingHistory.tsx';
-import { SelectMentor } from '@/components/ConsultingManagement/SelectMentor.tsx';
+import { ConsultingHistory } from '@/components/ConsultingManagement/ConsultingHistory';
+import { SelectMentor } from '@/components/ConsultingManagement/SelectMentor';
 import Community from '@/pages/Community';
 import { Consulting } from '@/pages/Consulting';
 import { ConsultingManagement } from '@/pages/ConsultingManagement';
